move base view model test to new file
